Fix 'learn more' link alignment in the bundle card.
