+++ client/jetpack-cloud/sections/partner-portal/license-bundle-card-description/style.scss
@@ -36,3 +36,8 @@
 .license-bundle-card-description__list li img {
 	margin-right: 0.75rem;
 }
+
+.license-bundle-card__more-info-link {
+	text-decoration: underline;
+	cursor: pointer;
+}
