+++ client/jetpack-cloud/sections/partner-portal/license-bundle-card/index.tsx
@@ -1,12 +1,8 @@
 import { Button } from '@automattic/components';
 import formatCurrency from '@automattic/format-currency';
 import { useTranslate } from 'i18n-calypso';
-import { useCallback, useState } from 'react';
-import { useDispatch } from 'react-redux';
 import LicenseBundleCardDescription from 'calypso/jetpack-cloud/sections/partner-portal/license-bundle-card-description';
-import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { APIProductFamilyProduct } from '../../../../state/partner-portal/types';
-import LicenseLightbox from '../license-lightbox';
 import { getProductTitle } from '../utils';
 
 import './style.scss';
@@ -21,27 +17,6 @@ export default function LicenseBundleCard( props: Props ) {
 	const { tabIndex, product, onSelectProduct } = props;
 	const productTitle = getProductTitle( product.name );
 	const translate = useTranslate();
-	const dispatch = useDispatch();
-	const [ showLightbox, setShowLightbox ] = useState( false );
-
-	const onShowLightbox = useCallback(
-		( e: any ) => {
-			e.stopPropagation();
-
-			dispatch(
-				recordTracksEvent( 'calypso_partner_portal_issue_license_product_view', {
-					product: product.slug,
-				} )
-			);
-
-			setShowLightbox( true );
-		},
-		[ dispatch, product ]
-	);
-
-	const onHideLightbox = useCallback( () => {
-		setShowLightbox( false );
-	}, [] );
 
 	const onSelect = () => {
 		onSelectProduct( product );
@@ -52,11 +27,7 @@ export default function LicenseBundleCard( props: Props ) {
 			<div className="license-bundle-card">
 				<div className="license-bundle-card__details">
 					<h3 className="license-bundle-card__title">{ productTitle }</h3>
-					<LicenseBundleCardDescription product={ product } />
-
-					<Button className="license-bundle-card__more-info-link" onClick={ onShowLightbox } plain>
-						{ translate( 'Learn more' ) }
-					</Button>
+					<LicenseBundleCardDescription product={ product } onSelect={ onSelect } />
 
 					<div className="license-bundle-card__pricing">
 						<div className="license-bundle-card__price">
@@ -77,15 +48,6 @@ export default function LicenseBundleCard( props: Props ) {
 					{ translate( 'Select' ) }
 				</Button>
 			</div>
-
-			{ showLightbox && (
-				<LicenseLightbox
-					ctaLabel={ translate( 'Select License' ) }
-					product={ product }
-					onActivate={ onSelect }
-					onClose={ onHideLightbox }
-				/>
-			) }
 		</>
 	);
 }
